patch: applied/page-zerox
add a zerox-like functionality accessible via the middle mouse button.

creates a new window with the current image only, useful for selecting important
pages of large documents.

note that it doesn't implement _full_ Zerox functionality because modifications
to the current image aren't reflected in the zerox-ed one. it is doubtful that
such functionality is needed.

man page will be modified as soon as the name 'zerox' is agreed upon.

Reference: /n/sources/patch/applied/page-zerox
Date: Wed Jun 16 19:13:52 CES 2004