patch: applied/_good_libcontrolfix
This patch fixes a silly bug in libcontrol.
It would complaint about column not being a control type.
There was an extra Ctltextbutton3 or similar that was set as nil
in the list of controls. So, _ctllookup would stop looking up
controls.
I changed _ctllookup not to stop on the first nil entry, to avoid this
kind of bug in the future.

I am sorry for the ugly name for the patch. I hit ESC while
sending the "libcontrolfix" patch but patch/create did send
the patch, so I couldnt recreate it with a non nil comment.

hth

Notes:
Thu Jul 1 22:21:45 EDT 2004 jmk
    i've asked sape to look a this because our local libcontrol
    source is newer than that on sources and maybe we should update
    sources.

    --jim

Reference: /n/sources/patch/applied/_good_libcontrolfix
Date: Fri Jul  2 18:15:48 CES 2004
Reviewed-by: jmk