patch: applied/sokoban-multistep
A multi-step-move change to the original Plan 9 sokoban.
The new behaviour is only visible in the Button 3 menu
(new entry [no]animate), and when you click button 1
on a square not adjacent to glenda - otherwise the
behaviour should be as before.

This is essentially the same change I already
reported to 9fans, only the code has been cleaned
up slightly.

Axel.

Reference: /n/sources/patch/applied/sokoban-multistep
Date: Tue Dec  7 18:04:51 CET 2004