patch: applied/pragincomplete
this is the patch i mentioned several weeks ago.
it extends #pragma incomplete to allow -T to be switched off
and on for whole libraries (eg, in the Freetype include files),
mainly intended for use when the code is imported
and so constructed that it is hard to add
all required `incomplete' names.

Reference: /n/sources/patch/applied/pragincomplete
Date: Wed Feb  2 18:24:06 CET 2005