patch: applied/jukefs-mapname
This is a fix of jukefs, where there could have happen a reread()
on a static buffer, which Plan9 does not allow.

This patch was done in a distributed manner on IRC, with some beer, red wine and across four countries.

Notes:
Fri Apr 8 11:02:38 EDT 2005 rsc
    This would have been a lot easier to follow with a
    better explanation.  reread() doesn't take any arguments
    at all.  The problem is that mapname, which might point at
    an argument or a static string, is assumed to be a 256-byte
    buffer, which it isn't.

    Thanks for the patch.

Reference: /n/sources/patch/applied/jukefs-mapname
Date: Wed Apr  6 20:32:36 CES 2005
Reviewed-by: rsc