patch: applied/install-mbr-question
This is a patch for adding a question to the user in
bootplan9, whether he wants to install the Plan9 MBR
or not. The existing mechanism only asked when the
Plan 9 partition was 2Gb into the disk. Now, when this
happens, some hints about it are printed additionally
to the main question.

Notes:
Mon Apr 11 16:44:39 EDT 2005 rsc
    There was an unmatched quote in one of the new sections.
    Please check that I got it right.  I removed the sectio
    with the bad quote -- it wasn't adding anything anyway.

    Russ

Reference: /n/sources/patch/applied/install-mbr-question
Date: Sat Apr  9 14:26:28 CES 2005
Reviewed-by: rsc