patch: applied/tar-seek
when printing a table of contents or skipping an archive member
because it didn't match any of the command-line patterns,
seek past large archive members when the archive is on a seekable
medium.  this is a big time-saver after archiving large files
(such as other tar archives), when you use `tar t' to generate
a table as contents as quick verification that everything got
written into the archive.

Reference: /n/sources/patch/applied/tar-seek
Date: Fri Apr 15 09:25:47 CES 2005