patch: sorry/bargraph-wsys
This patch fixes the problem that although the user did
specify textinst for the installation and bargraph gets
the parameter -t, the warning about the undefined $wsys
is printed out. This happened because of a OR in main()
of bargraph, where newwin() is executed, but should not.

Notes:
Sun Apr 24 13:47:05 EDT 2005 rsc

    Ignoring the setting of textmode and rbar,
    the old code said

    	if(textmode || newwin(p) < 0)
    		;
    	else
    		initdraw();

    while the new code says

    	if(!textmode){
    		if(newwin(p) < 0)
    			;
    		else
    			initdraw();
    	}

    and I fail to see how these are different.
    I'm sure it's something obvious -- please help.

Reference: /n/sources/patch/sorry/bargraph-wsys
Date: Sun Apr 24 19:02:02 CES 2005
Reviewed-by: rsc