patch: sorry/earliest-received
-- Old code if'ed out, should be removed --

Originally, the first "Received:" header found was used to determine
a "reliable" date.  In my understanding, this ought to be the "last"
one, as that is the "earliest" stamp and ought to be the one that
best matches the requirements.

Mbox.c adjusted to meet my understanding.  It seems to work reliably,
but further tests are required.

Notes:
Sun May 1 13:06:34 EDT 2005 rsc
    This patch isn't correct.  The convention is that the message date
    is the arrival time not the sending time.  In addition to breaking
    this convention, taking the innermost received means we have to
    figure out how to convert from that guy's time zone to our own,
    as it says in the comment.

    By the way, your patch also seemed to delete support for
    windows-1251 and koi8-r.  Have you pulled recently?

    Russ

Reference: /n/sources/patch/sorry/earliest-received
Date: Sun May  1 07:54:17 CES 2005
Reviewed-by: rsc