patch: applied/exportfs-abugs
This is a patch for exportfs and its manpage. It fixes the issue in
exportfs, that if you specify -n (no none) or -N (other namespace)
after -a (authenticate), that these options have no effect. This was
due to the fact that -a was handled in and not after the argument
processing. In the manpage the wrong stated '-n' is corrected to be
the real '-N' and '-n' is now documented.

Notes:
Thu May 5 13:31:13 EDT 2005 rsc
    I fixed this, and also the opening of srvfd.
    Your code looked fine except for the use of 'y' and 'n'
    as flag values -- in Plan 9 we tend to use a simple int
    with 0 or 1.

    I also updated the man pages to describe -B too.

Reference: /n/sources/patch/applied/exportfs-abugs
Date: Thu May  5 18:27:10 CES 2005
Reviewed-by: rsc