patch: sorry/authsrv-keyfs
This patch adds a check to authsrv, if keyfs is running or
not and then starts it. This will solve some annoying sit-
uations, where you play with Listen and forget to start Li-
sten in a new namespace on a remote machine that is hund-
rets of kilometers away.

Notes:
Sun May 8 12:42:40 EDT 2005 rsc
    I'm not so sure about this one.  This would have multiple keyfs
    running at the same time, and I'm not sure that they lock between
    themselves properly.  Better to start it correctly to begin with.

Reference: /n/sources/patch/sorry/authsrv-keyfs
Date: Fri May  6 14:54:26 CES 2005
Reviewed-by: rsc