patch: applied/page-pdf-crop
Correct problem with page(1) reported by Peter Cejchan, which turned out
to be incorrect handling of (rare) PDF files with a CropBox inset within
the MediaBox.  The accompanying demo file croptest.ps (which I don't
expect you to copy to sources) illustrates the effect: it prints text
in four orientations, translated to each quadrant of the page - 16 pages
in all.  When translated with ps2pdf and viewed with page, the result
should be cropped to the correct quadrant and displayed with text upright.

Notes:
Sat May 14 12:01:20 EDT 2005 rsc
    Thanks for tracking this down.

Reference: /n/sources/patch/applied/page-pdf-crop
Date: Fri May 13 12:30:40 CES 2005
Reviewed-by: rsc