patch: sorry/upper_etheraddr
It's hard to believe that somebody added the ability to deal with
colons, but not upper case.  On the other hand, maybe strtoul() is
to blame for not understanding uppercase HEX digits.  In which
case, somebody ought to look into it.  What do the standards say?

Notes:
Sat May 14 12:02:16 EDT 2005 rsc
    Please send a test program that demonstrates
    what you are trying to fix.  Even better, also
    explain what real program this helps.  I'm just
    confused.

    Thanks.

Sat May 14 12:04:16 EDT 2005 rsc
    Actually, having read the 9fans conversation, I think
    this is just misguided.  Ndb doesn't even use parseether.

Reference: /n/sources/patch/sorry/upper_etheraddr
Date: Sat May 14 08:35:54 CES 2005
Reviewed-by: rsc