patch: applied/install-spelling
This patches fixes some bad style sentences, an error in "halt",
because of kfs, some bad spelled sentences and the GB form of
"Initializing" in disk/format.

Thanks to Vester Thacker <vester.thacker@gmail.com> for pointing
this out.

Still one problem is there: the description of mountfs, stop and
finish are placed wrong in the menus because of their short names.
I don't see a real solution for this.

Reference: /n/sources/patch/applied/install-spelling
Date: Wed May 18 20:14:55 CES 2005