patch: applied/hget-setoffset-postbody
Change usage string so that it matches the hget(1) manual page.

Move the printing of u->postbody after the printing of cookies, since
cookies must appear in the header.

Add indentation to hhcrange().

Change setoffset() so that it sets out->written, and seeks to the
appropriate file offset if the -o option is used.  Without this, the
assert(d >= 0) in output() fails and the output becomes garbage if the
file was already partially downloaded.

Reference: /n/sources/patch/applied/hget-setoffset-postbody
Date: Wed Jun 15 12:47:25 CES 2005