patch: applied/4s-leftborder-whitening
fix the problem where the long narrow yellow brick in vertical
position erases the left border.

It did this, because it was defined as a 2 column 4 row high brick,
where the leftmost column was entirely empty (i.e. white).

I have no clue why the brick was defined this way.
Redefining the brick to be single column 4 row high seems to
work and not have hidden repercussions.

There may be a deeper fix that ``ignores'' the entirely white
column (or masks it away), but I leave that for a more expert fixer.

Axel.

Reference: /n/sources/patch/applied/4s-leftborder-whitening
Date: Wed Jun 15 16:20:18 CES 2005