patch: sorry/rio-piemenu-update
This patch is a small change to the saved patch "piemenus-rio".
It changes the behaviour of the piemenus, to be no more drawn
outside of the screen. Besides this is there a style update, to
be more consistent with the "Plan 9 style option int variables"
and the style of whole libdraw.

Reference: /n/sources/patch/sorry/rio-piemenu-update
Date: Tue Aug  9 10:57:30 CES 2005