patch: applied/dns-soa-expire
As DeNIC wants to have other values, then what is default and
not changeable in /lib/ndb/local, was this patch created. Now
you can change the SOA values of "retry" and "expire", where
the last one is expected to be higher than 604800, when host-
ing a .de domain - in contrast to the 86400 (Day) default.

Reference: /n/sources/patch/applied/dns-soa-expire
Date: Tue Aug  9 17:09:25 CES 2005