patch: applied/acme-enable-menu-via-ctl
interpret menu and nomenu in the ctl file to set resp. unset filemenu
which controls whether or not the automagic Undo, Redo, Put are in the tag.

this is needed to allow automagic Undo/Redo/Put for acme Wiki --
because all files edited in the wiki look like directories to acme
the automagic Undo/Redo/Put are switched off there.

a separate patch (acme-wiki-filemenu-in-tag) depends on this one.

Axel.

Reference: /n/sources/patch/applied/acme-enable-menu-via-ctl
Date: Mon Aug 15 00:55:24 CES 2005