patch: applied/acme-wiki-use-winismodified
use the acme ctl file ismodified field instead of
maintaining our own dirtied variable.
When the user tries to Get a modified file,
we do remember if we have warned already,
and reset that information in the same places
where we used to set the dirtied variable.

this patch includes the acme-wiki-filemenu-in-tag patch
(the change in wiki.c around 441c445,447) which depends
on acme-enable-menu-via-ctl , and the
acme-wiki-fix-dirtystate patch (ignore writes to body
or tag; winclean after setting file name and tag commands).

Axel.

Reference: /n/sources/patch/applied/acme-wiki-use-winismodified
Date: Mon Aug 15 14:58:49 CES 2005