patch: applied/spin425
Upgrade the version of Spin included with Plan 9 to 4.2.5.  NOTE: I do not use Spin, so I'm not sure it's 100% correct, but it seems to work with casual testing.

I also did not update the Spin documents in /sys/doc because I am unsure of their origin.

Reference: /n/sources/patch/applied/spin425
Date: Wed Aug 31 04:37:34 CES 2005