patch: applied/mv-remove
I tracked down an anoyance in my cifs client -it
will not let me delete open files - which I should fix.
The only cluprit I found was mv(1) and there seems to be no
reason for it do do this; so I "fixed" it.

I will not be surprised if this patch gets sorry'ed.

PS: tried to preserve the codeing conventions used in mv(1) but
they appear to be mixed.

-Steve

Notes:
Fri Sep 9 11:00:56 EDT 2005 rsc
    I cleaned up the code a little, but I kept the close before
    remove for you.  Do fix cifs, though.

Reference: /n/sources/patch/applied/mv-remove
Date: Thu Sep  8 16:13:57 CES 2005
Reviewed-by: rsc