patch: applied/httpd-range-compliance
This patch solves some problems in wget and Firefox, when doing a partial
request (Range:) on a file on a Plan 9 httpd. The Plan 9 httpd does not
send out a Content-Length: when doing so, but these applications expect to
have one and show it directly to the user.

One clause in the HTTP 1.1 RFC says in section 4.4 Message Length:
"For compatibility with HTTP/1.0 applications, HTTP/1.1 requests contain-
ing a message-body MUST include a valid Content-Length header field unless
the server is known to be HTTP/1.1 compliant."

Compliant or not complianet - here is the patch.

Notes:
Thu Oct 13 07:24:23 EDT 2005 rsc
    Installed and running on the plan9.bell-labs.com web server.
    Please be more careful about spacing issues -- make the
    spacing on changed code look like the spacing around it.

    Thanks.

Reference: /n/sources/patch/applied/httpd-range-compliance
Date: Thu Oct 13 09:40:19 CES 2005
Reviewed-by: rsc