patch: applied/lib9p
In file.c there seemed to be a leak. Also, the order of sons in a file
is preserved wrt creation order. If we create ./a ./b then rm ./b and
create ./c, the list of sons has ./a ./c.
This is important for omero, where the order of files determines the order
of widgets.

In srv.c an incref (already marked in sources(?)) was missing.

Notes:
Sat Oct 29 10:55:15 EDT 2005 rsc
    I applied your fix to srv.c, which was correct.

    I did not apply your change to removefile.  The reason
    the filelist structure exists at all is to keep them around
    beyond remove time because they are referenced by open
    directory readers (see bottom of file).  I made a different
    change that should have the effect you wanted - the listing
    of files in a directory reflects the create order.

    I could not tell from your diffs what the leak was
    in file.c.  Could you point it out to me?

    Thanks.
    Russ

Reference: /n/sources/patch/applied/lib9p
Date: Thu Oct 27 20:51:56 CES 2005
Reviewed-by: rsc