patch: sorry/rfc822date
Acme/Mail displays the date from the last Received: header, which
can be considerably different from the date when the message was
first sent.  The format of the acme/mail date is by its nature
local time, which is helpful.

This is a stab at displaying a more useful date, as in my particular
case messages are batched together and the date of delivery conveys
a false impression.  I have not felt a need to convert the date to
localtime, although no doubt that would appeal to many users.  It
is also not a trivial operation, of course, what with all the
formats in use.

I'm submitting this more for discussion than for actual incorporation,
although it would be nice not to have to treat this module as a
special case on replication.

Lucio De Re.

Notes:
Sun Nov 6 09:48:26 EST 2005 rsc
    As I'm sure you expected, I'm not going to apply this.

    You have correctly listed all the reasons that we use
    the received: line - it's local time, it reflects actual
    delivery, it's standardized and easy to parse (since we
    generated it), and so on.  It's also consistent with upas/fs.
    As far as parsing the RFC822 date goes, you can just copy
    /sys/src/cmd/upas/fs/strtotm.c.

    The error() changes are just wrong - the Mail: prefix is
    added by error and need not be in the string.

    A (clumsy) alternative is to read your mail via POP3, so
    that the Unix from line is dropped completely anyway.
    In this case upas/fs will use the RFC822 date because that
    is the only choice it has.

    Russ

Reference: /n/sources/patch/sorry/rfc822date
Date: Sat Nov  5 17:58:02 CET 2005
Reviewed-by: rsc