patch: sorry/plumbcreate
Change to plumbopen so that it tries to create the port file if it does
not exist. This should be a "nop" when using plumber, but it's used when
using youps (a file system that implements ports and + a cut-down plumber
that implements just the send and rules files).

Notes:
Fri Nov 18 11:28:23 EST 2005 rsc
    I'm not convinced.  The function is called plumbopen not plumbcreate.
    If you really want ports to appear on demand, why not have youps
    just create them as needed, like mntgen does.

Reference: /n/sources/patch/sorry/plumbcreate
Date: Fri Nov 18 10:48:43 CET 2005
Reviewed-by: rsc