patch: applied/namefixandleak
Avoids a leak of uid, and changes the default name for the root file of the
tree to "dev", because "/" is an illegal name that may lead to errors for
file system multiplexors.
This patch is not against the patched version of lib9p I'm using, it's against
the one in sources.

Notes:
Fri Nov 18 11:41:56 EST 2005 rsc
    Applied the leak fix.  Not convinced about the name.
    I think it's perfectly sensical for a stat of the root
    to give "/" as the name, and I can't find it written
    anywhere that that's an illegal name.

    In fact, if you ls -ld /, that's the name you get.

Reference: /n/sources/patch/applied/namefixandleak
Date: Fri Nov 18 11:11:10 CET 2005
Reviewed-by: rsc