patch: applied/wikifs-not-allow-questionmark-in-title
Pages with '?' in their name break links in the web interface.

The proper solution is maybe to generate links that are
properly url-encoded, that would remove the limitations on <>
too, but url-encoding is trickier than it seems.

But for now it's easier to just keep people from creating
such pages.

Reference: /n/sources/patch/applied/wikifs-not-allow-questionmark-in-title
Date: Tue Dec 13 09:42:11 CET 2005