patch: applied/usbd-vid-did
Adds vid and did output to usb/usbd. First, this is required because of the
USB spec, that says, that the USB software should first look for vid and did
and then for the device class. Second, for example an UART driver has no USB
device class and needs checking of the vid and did.

Reference: /n/sources/patch/applied/usbd-vid-did
Date: Thu Jan 19 16:18:09 CET 2006