patch: applied/9fs-patch-anon
A patch to fix replica/pull ... network to work with anonymous sources and add
options to 9fs, that make it possible for new users to access 9fs. The solution
to have a second case was choosen, so there won't be irritations in the future,
about; if we are now using sources anonymiously or not.

Notes:
Thu Feb 9 10:40:45 EST 2006 rsc
    Thanks for pointing out that /dist/replica/network
    wasn't doing the right thing.  I've fixed it to use 9fs sources.

    I do not want 9fs anonsources -- the way to access sources
    has always been "9fs sources" and I do not want that to change.
    I did clean up case sources in /rc/bin/9fs though.

    Russ

Reference: /n/sources/patch/applied/9fs-patch-anon
Date: Wed Feb  8 10:59:27 CET 2006
Reviewed-by: rsc