patch: applied/iostats
a few minor problems:
- modes weren't printed properly (fcallfmt assumes dirmodefmt and dirfmt).

- the number of workbufs wasn't sufficient for lots of parallel reads (e.g. when testing fcp)
the fix should probably allow an indefinite number, but 40's better than 16!

- it didn't follow $path when executing a file. i'm aware that $path's not
so important in plan 9, but i (and perhaps others too) have path=(/bin .)
and it can lead to surprises when iostats tries things the other way around.
it's a pity that rdenv() is so painful, but i'm not aware of an alternative.

Reference: /n/sources/patch/applied/iostats
Date: Wed Feb 15 14:56:29 CET 2006