patch: applied/auth-stderr-tags
I've added a tag with the program's name to each fprint(2, ...)
in /sys/src/cmd/auth that wasn't obviously debugging output or
part of a protocol ('!' seems to be magic at the start of some
messages).  In some cases, I switched to using sysfatal().

It all compiles and I'm about to install it and reboot the world.

Reference: /n/sources/patch/applied/auth-stderr-tags
Date: Sat Feb 18 05:38:47 CET 2006