patch: applied/kmem-man-index
Add kmem to 1/INDEX and under NAME in leak(1).

And *PLEASE* could we have the lookman index re-built daily? it takes just a few seconds to rebuild, and it is a pain in the ass when looking for things and you can't find them because turns out the indexes are out of date.

Notes:
Mon Apr 10 12:09:16 EDT 2006 rsc
    please don't submit patches for mechanically generated files.  thanks.

Reference: /n/sources/patch/applied/kmem-man-index
Date: Sat Apr  8 23:42:48 CES 2006
Signed-off-by: uriel@cat-v.org
Reviewed-by: rsc