patch: sorry/acme4-event-description-clarification
Adds a bit of text to explain that the "event" file for acme doesn't take the same format for input
written to the file as it outputs.

Based on the function I posted to 9fans a few days ago for "winclear".

Notes:
Sat Aug 12 02:34:38 EDT 2006 rsc
    I think that section is perfectly clear.
    It's only talking about the so-called output messages.
    (One man's output is another man's input, of course.)
    At the end of the section it does talk about which
    messages can be written back, and what format they must take.

Reference: /n/sources/patch/sorry/acme4-event-description-clarification
Date: Mon Jul 17 08:20:57 CES 2006
Signed-off-by: leimy@plan9.jp
Reviewed-by: rsc