patch: applied/test-invalid-args-and-man-page
Make test(1) complain when we pass unexpected extra args
instead of silently ignoring them or returning TRUE.
This is a bit more consistent with PoSix systems and
avoids confusion (fgb thought that test -dw /tmp would
check if /tmp was a writable dir).

Also complain when the argument to -t is not a valid
filedes.

Fix small typo in the man page and document that there
is on way to test if a dir is writeable.

uriel

Notes:
Sat Aug 12 01:57:48 EDT 2006 rsc
    I don't think it's a bug that -w tells you
    that directories cannot be opened OWRITE.

Reference: /n/sources/patch/applied/test-invalid-args-and-man-page
Date: Mon Jul 24 22:20:54 CES 2006
Signed-off-by: uriel@cat-v.org
Reviewed-by: rsc