patch: applied/respondcmderror-install-quotefmt
Adding a quotefmtinstall call to respondcmderror (the way it is done
in libauth too), since it is using a "%q" verb.

Before that I got messages like

	unknown control message: "%q% %q% %q% %q%"

when trying to partition a usbsfs data file, since usbsfs apparently
doesn't know the `part' command.

With the patch it prints
	unknown control message: "part 1 2 33"

Reference: /n/sources/patch/applied/respondcmderror-install-quotefmt
Date: Thu Sep 14 18:38:11 CES 2006
Signed-off-by: mt@wmipf.in-berlin.de