patch: sorry/smtpd_quotes-in-addresses
There are possible relaying problems with quotes in rcpt addresses.
We talked about stripping quotes or refusing addresses with quotes in
smtpd.

This change adds '"' to the list of illegal characters and makes smtpd
refuse such addresses.  (shellcars() returns 1 when verifying sender
and rcpt addresses.)

Let me know if that's not the right way to go and you rather want to
see quotes stripped than rejected.

Notes:
I think we can't just reject quotes; they are permitted
in the local-part of an address.  We can reject an address
that starts and ends with quotes, however, since the domain
part cannot be quoted.
Thu Oct 19 21:31:21 EDT 2006 geoff
    We seem to be in agreement that the fix is in rewrite.
    New rewrite files with the fixes were just pushed out.

Reference: /n/sources/patch/sorry/smtpd_quotes-in-addresses
Date: Wed Oct 18 15:56:48 CES 2006
Signed-off-by: hdd@voidness.de
Reviewed-by: geoff