patch: sorry/udp-manpage-fix
Updates ip(3) man page to correctly describe the header-format
commands which may be sent to the ctl file for UDP and RUDP.
In passing, fixes a typo in ip.h.

There is an argument for omitting description of the "oldheaders"
format of UDP, but it appears to correspond to the "headers"
format of RUDP(!?) so I described it.

Notes:
Fri Oct 20 17:50:23 EDT 2006 geoff
    As we discussed, this is based on an out-of-date page.

Reference: /n/sources/patch/sorry/udp-manpage-fix
Date: Thu Oct 19 07:03:27 CES 2006
Signed-off-by: davide+p9@cs.cmu.edu
Reviewed-by: geoff