patch: applied/tr2post-creator
Adds the creator to PostScript files generated by tr2post.
When/if the PostScript program is distilled into a PDF file,
anyone who cares to look up the application in Adobe
Acrobat Reader (or whatever) will see that Plan 9 was
involved in the files creation! (not to mention troff)
--btdn

Reference: /n/sources/patch/applied/tr2post-creator
Date: Fri Dec  1 00:32:15 CET 2006